Introduces the idea of profiles
Profiles are added in the 'profiles' directory.

The install script runs through each directory and runs 'run_profile' to determine if the profile should be executed or not.

If so, dotbot is run for that profile.

The idea is to add more flexibility; e.g, we can now introduce more specific profiles (like Ruby, golang etc).